move misplaced table-only filters to TableFilterBuilder
